Moves container into page
